Get started with socket.io
sec. 17, video 154

$ npm i socket.io@2.2.0

Create the server outside of the Express library and configure
it to use our Express app.
Initialize a new instance of socket.io by passing the HTTP server
object.
When Express creates the server behind the scenes, we don't
have access to it to pass it into socketio().